+++ src/static-pages/components/EditorialStaff.tsx
